fix(search): anchor past a row's kind decoration; scope the substring property
CI caught what my local run's random inputs missed: the proptest
`substring_matcher_agrees_with_lowercase_contains` asserts that any query
without `*?[` behaves as a substring, and `^`/`$` grew the metacharacter set
past the three its strategy excludes (minimal case: name="", query="^").
Scope the property to its actual subject with a prop_assume, and add
`anchors_agree_with_starts_and_ends_with` so the anchors get differential
coverage of their own against starts_with/ends_with/eq.

Chasing it surfaced a real defect. Rows are matched by their DISPLAY name, and
`Entry::display_name` appends `/` to a directory and `*` to an executable — so
a naive end anchor could never match either: `/src$` would miss `src/` and
`/sh$` would miss `deploy.sh*`. That is the same silent no-match that made `^`
worth fixing. Anchored queries now get their own `Matcher::AnchoredGlob`, which
retries against the undecorated name; plain globs are untouched, so `=*/`
still means "directories only".

`Matcher` moves to `src/app/matcher.rs` — the additions pushed mod.rs past the
1500-line anti-monolith ceiling, and it's a pure text leaf with no `App`
dependency, so it extracts rather than the ceiling moving.
