+++ AGENTS.md
@@ -49,7 +49,8 @@ Deep design decisions live in [`ARCHITECTURE.md`](ARCHITECTURE.md) (sync-only, M
 
 **`src/app/`** — the application layer (MVU). `App` owns three disjoint fields: `state: AppState` (Model — pure domain, no OS handles), `runtime: Runtime` (OS handles/channels/PtyHosts), `view: ViewState` (render ephemerals/caches). New handler/render/command/action logic goes in the matching child module below — **never back into `mod.rs`** (ceiling-guarded). Child modules read `App`'s private fields via the descendant-module rule.
 
-- **`mod.rs`** — module root: the three struct defs + `Message` enum + small glue (`sh_c`, `row_from_entry`, `Matcher`, `open_help`).
+- **`mod.rs`** — module root: the three struct defs + `Message` enum + small glue (`sh_c`, `row_from_entry`, `open_help`).
+- **`matcher.rs`** — the `Matcher` behind `/` search and the `=` limit filter: case-insensitive substring, glob on `*`/`?`/`[`, and `^`/`$` anchors desugared into the glob engine. An `AnchoredGlob` also matches past a row's kind decoration (`/` dir, `*` exe) so an end anchor can find `src/`. Pure text, no `App` dependency.
 - Loop core — **`run.rs`** (`App::run`, dispatch/render/teardown), **`bootstrap.rs`** (`App::new`), **`proc.rs`** (input-reader thread + foreground-exec runner), **`update.rs`** (`App::update`, the single update entry), **`util.rs`** (App-independent leaf helpers).
 - **`state/`** — `AppState`, the Model. Per-column browser fields (`listing`, `cursor`, `rows`, `picks`, `masks`, `temp_filter`, `view`, sort, `list_generation`) bundle into a `Commander`; `left` always present, `right: Option<Commander>` for the second column. The update path reaches the focused column via `cur()`/`cur_mut()`; render addresses `left`/`right` explicitly. App-wide state (marks, inventory, graveyard, harpoon, pane, focus, config, mode, flash, vsplit, git) stays flat. `impl` splits by domain (`navigation`/`selection`/`listing`/`git`/`apply`/`dispatch`); `deny(unwrap_used)`.
 - **`render/`** — the View (pure `&self`); `deny(unwrap_used)`. `mod` = frame lifecycle (`compute_layout`, `render`, `prepare_*`); `inner`/`chrome`/`overlays` paint.
